SanitiseTransform: Create "housekeeping" transformation for scheduler
This so far includes resolving associates and sequence associations,
but could be extended into a general-purpose input unification pipeline.
