+++ loki/transform/transform_inline.py
@@ -27,8 +27,6 @@ from loki.subroutine import Subroutine
 
 from loki.transform.transformation import Transformation
 from loki.transform.transform_dead_code import dead_code_elimination
-from loki.transform.transform_sequence_association import transform_sequence_association
-from loki.transform.transform_associates import resolve_associates
 from loki.transform.transform_utilities import (
     single_variable_declaration,
     recursive_expression_map_update
@@ -60,11 +58,6 @@ class InlineTransformation(Transformation):
     inline_marked : bool
         Inline :any:`Subroutine` objects marked by pragma annotations;
         default: True.
-    resolve_associate_mappings : bool
-        Resolve ASSOCIATE mappings in body of processed subroutines; default: True.
-    resolve_sequence_association : bool
-        Replace scalars that are passed to array arguments with array
-        ranges; default: False.
     eliminate_dead_code : bool
         Perform dead code elimination, where unreachable branches are
         trimmed from the code; default@ True
@@ -84,7 +77,6 @@ class InlineTransformation(Transformation):
     def __init__(
             self, inline_constants=False, inline_elementals=True,
             inline_internals=False, inline_marked=True,
-            resolve_associate_mappings=True, resolve_sequence_association=False,
             eliminate_dead_code=True, allowed_aliases=None,
             remove_imports=True, external_only=True
     ):
@@ -93,8 +85,6 @@ class InlineTransformation(Transformation):
         self.inline_internals = inline_internals
         self.inline_marked = inline_marked
 
-        self.resolve_associate_mappings = resolve_associate_mappings
-        self.resolve_sequence_association = resolve_sequence_association
         self.eliminate_dead_code = eliminate_dead_code
 
         self.allowed_aliases = allowed_aliases
@@ -102,15 +92,6 @@ class InlineTransformation(Transformation):
         self.external_only = external_only
 
     def transform_subroutine(self, routine, **kwargs):
-
-        # Associates at the highest level, so they don't interfere
-        # with the sections we need to do for detecting subroutine calls
-        if self.resolve_associate_mappings:
-            resolve_associates(routine)
-
-        # Transform arrays passed with scalar syntax to array syntax
-        if self.resolve_sequence_association:
-            transform_sequence_association(routine)
 
         # Replace constant parameter variables with explicit values
         if self.inline_constants:
